+++ e2e/cli_test.go
@@ -92,10 +92,9 @@ func Test_E2E_CLI(t *testing.T) {
 					// Check log file
 					logData, _ := utils.ReadFixture(tt.Args.ExpectedLog.LogFile, "output")
 					validation := tt.Args.ExpectedLog.ValidationFunc(logData)
-					/*require.Truef(t, validation, "The output log file 'output/%s' doesn't match the regex validation",
-					tt.Args.ExpectedLog.LogFile)*/
-					require.Truef(t, validation, "olha aqui %s",
-						logData)
+
+					require.Truef(t, validation, "The output log file 'output/%s' doesn't match the regex validation",
+						tt.Args.ExpectedLog.LogFile)
 				}
 
 				if tt.Args.ExpectedOut != nil {
@@ -107,12 +106,9 @@ func Test_E2E_CLI(t *testing.T) {
 					formattedWant := loadTemplates(want, templates)
 
 					// Check number of Lines
-					/*require.Equal(t, len(formattedWant), len(out.Output),
-					"[fixtures/%s] Expected number lines: %d\n[CLI] Actual KICS output lines: %d",
-					tt.Args.ExpectedOut[arg], len(formattedWant), len(out.Output))*/
 					require.Equal(t, len(formattedWant), len(out.Output),
-						"o output %S",
-						out.Output)
+						"[fixtures/%s] Expected number lines: %d\n[CLI] Actual KICS output lines: %d",
+						tt.Args.ExpectedOut[arg], len(formattedWant), len(out.Output))
 
 					// Check output lines
 					for idx := range formattedWant {
